Update README.md
hi! I have added a few more links to the resources that I usually follow and were missing here i.e. I have worked on the issue#17 Add more awesome links. 
Please give it a check and let me know if my contribution was worthy or not! thanks in advance.